Move types/configs out of 'Lightning.ts'
