+++ src/Lightning.ts
@@ -1,7 +1,6 @@
 import assert from "assert"
 import { createHash } from "crypto"
 import {
-  AuthenticatedLnd,
   cancelHodlInvoice,
   createInvoice,
   getInvoice,
