+++ src/LightningTypes.d.ts
@@ -1,3 +1,5 @@
+type AuthenticatedLnd = import("lightning").AuthenticatedLnd
+
 interface IAddInvoiceRequest {
   value: number
   memo: string | undefined
